+++ sortings/sortings/func.h
@@ -11,10 +11,10 @@
 
 #include <stdio.h>
 #include <IOStream>
-void quickSort(int arr[], int left, int right,int arr2[]);
-void bubbleSort(int arr[],int arr2[],int n);
+void quickSort(int arr[], int left, int right,double arr2[]);
+void bubbleSort(int arr[],double arr2[],int n);
 void showArray(int arr[],int n);
 void createArray(int arr[],int n);
 void copyArray (int arr[],int n, int arr2[]);
-void wtfSort(int arr[],int n, int arr1[],int arr2[]);
+void wtfSort(int arr[],int n, int arr1[],double arr2[]);
 #endif /* defined(__sortings__func__) */
